feat: add other api

+++ src/users/users.controller.ts
@@ -1,24 +1,26 @@
 import {
   Controller,
   Get,
-  Post,
   Body,
   Patch,
   Param,
   Delete,
+  UseGuards,
+  Request,
+  HttpStatus,
+  HttpException,
+  Req,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
-import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
+import { AuthGuard } from '@nestjs/passport';
+import { RolesGuard } from 'src/role/role.guard';
+import { Roles } from 'src/role/roles.decorator';
+import { Role } from 'src/role/utils/roles.enum';
 
 @Controller('users')
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
-
-  @Post()
-  create(@Body() createUserDto: CreateUserDto) {
-    return this.usersService.create(createUserDto);
-  }
 
   @Get()
   findAll() {
@@ -30,13 +32,27 @@ export class UsersController {
     return this.usersService.findOne(+id);
   }
 
+  @Roles(Role.Admin)
+  @UseGuards(RolesGuard)
+  @UseGuards(AuthGuard('jwt'))
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
-    return this.usersService.update(+id, updateUserDto);
+  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto, @Req() request) {
+    if(id == request.user.id) {
+      return this.usersService.update(+id, updateUserDto);
+    } else {
+      throw new HttpException('FORBIDDEN', HttpStatus.FORBIDDEN);
+    }
   }
 
+  @Roles(Role.Admin)
+  @UseGuards(RolesGuard)
+  @UseGuards(AuthGuard('jwt'))
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.usersService.remove(+id);
+  remove(@Param('id') id: string, @Request() request) {
+    if(id == request.user.id) {
+      return this.usersService.remove(+id);
+    } else {
+      throw new HttpException('FORBIDDEN', HttpStatus.FORBIDDEN);
+    }
   }
 }
